backend/requirements.txt
# FastAPI and Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Configuration
python-dotenv==1.0.0

# Redis
redis==5.0.1

# File handling
aiofiles==23.2.1

# HTTP Client
httpx==0.26.0

# Testing
pytest==7.4.0
pytest-asyncio==0.23.0
